requirements.txt
# 核心数据处理
pandas>=1.5.0
numpy>=1.24.0
tushare>=1.2.89

# Web框架
fastapi>=0.104.0
uvicorn>=0.24.0

# 配置文件
pyyaml>=6.0

# HTTP请求
requests>=2.31.0